// src/store/question/Types.ts
export interface QuestionTypes {
    acsOrder?: boolean,
    recordPage?:any,
    academic_year?:any,
    search:string,
    grade?:string,
    standard?:string,
    hasMore:boolean,
    SortOrderData:string,
    OrderNameData:string,
    page?:any,
    page_no?:any
}